small fixes
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@16471 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/apps/soundrecorder/DrawingTidbits.cpp
@@ -67,7 +67,7 @@ ReplaceTransparentColor(BBitmap *bitmap, rgb_color with)
 	ASSERT(bitmap->ColorSpace() == B_COLOR_8_BIT); // other color spaces not implemented yet
 	
 	BScreen screen(B_MAIN_SCREEN_ID);
-	uint32 withIndex = screen.IndexForColor(with); 
+	uint8 withIndex = screen.IndexForColor(with); 
 	
 	uchar *bits = (uchar *)bitmap->Bits();
 	int32 bitsLength = bitmap->BitsLength();
